Incorporate nic_ui into lwaftr run.lua; use CSV stats.
This patch incorporates the guts of the lwaftr "nic_ui" script into
the lwaftr's program script, with the goal of removing nic_ui later
and avoiding having to document it for now.  It also uses a new
CSV statistics module instead of the Statistics element that we
added last week.  The CSV stats module uses timers and counters to
get its information instead of apps, which should result in a lower
overhead.