Avoid constants.ethernet_header_size in fragment_ipv4()
This makes fragment_ipv4() receive the size of the L2 headers as a parameter,
instead of relying on constants.ethernet_header_size being updated e.g. when
VLAN tagging is enabled. This allows using the function to prepare packets
which then will be used to test the IPv4 reassembly code.

This change is a backporte of the relevant parts of commit 6a69fdf in the
lwaftr_olive branch.