MTU settings are now L3 rather than L2 based.
IE, an IPv6 MTU of 1500 is now 1500 bytes for the IPv6 header and payload,
rather than 1500 - (ethernet + possibly vlan) bytes.
All tests are updated accordingly and pass.
This also corrects the result packet for replying to attempt to fragment a
DF packet with a vlan: it's now 576 rather than 580 bytes, in accordance
with specifications on ICMP packet size.