Add range map
A range map divides the space of uint32 values into partitions that map
to the same value.  We will use the range map for lookup of the A/K
parameters that compute PSID, allowing the lwAFTR to efficiently support
a variety of A/K values within one binding table.

A simple test loop, including a fair bit of int/float conversion on the
LuaJIT side, results in 23 cycles / 6.77 nanoseconds per lookup on this
old i7-3770 sandy bridge desktop machine.  This overhead should be
constant, considering that we do lookup via an optimized branchless
binary search.