Incorporate new binding table.
This is a big ol' commit!  It does many things at once, and in fact
un-does some things.

Firstly, it switches the binding table to use the new parser.  It also
changes the syntax of the binding table.  See the README for
compile_binding_table for more details.  It re-incoporates the address
map into the binding table, because we don't want to run the risk that
there is somehow skew between the binding table at time T and the
address table at time T' -- the need to be all compiled together and,
when loaded in compiled form, all loaded together.

The binding table is compiled automatically and as needed.  We can also
compile the binding table ahead of time, using the compile-binding-table
command (which replaces compile-address-map).

The default name for the binding table in our configs is now
binding-table.txt.  By default compilation will strip a .txt suffix and
replace it with .o.

The binding table has been integrated into the lwaftr, but there are
things that can be cleaned up there.  In particular the clownshoes about
whether an IPv4 address is a big-endian integer, a little-endian
integer, or a uint8_t[4] is a bit irritating.  Dunno.  I'm feeling
pretty good about this!

We also need to change the PSID info map (previously called the address
map) to return 0 for the PSID parameters when an IPv4 address is not in
the map.  This will allow for cheap hairpinning checks instead of doing
a full binding table lookup.

Performance-wise we seem to have a regression, at least from what I can
tell locally -- going down from 18/16 Gbps in the benchmark result to
something more like 17/15 Gbps.  I hope that we can speed this up,
firstly with this address map change, and thereafter with other fixes
like removing C.foo calls from hot paths.