Improve lwaftr config parser error messages
Given the following input in a configuration file for the lwaftr:

   aftr_ipv4_ip = "1.2.3.4"

The following error would be generated when parsing it:

  lib/protocol/ipv4.lua:79: variable 'address' is not declared

...which is confusing, because the error does not point to the configuration
file, and does not point out the actual problem. With this patch applied the
message is instead:

  apps/lwaftr/conf_parser.lua:28: lwaftr.conf:1:15: error: IPv4 address expected

...which points out exactly the problem, and what the parser was expecting
to find. This patch also modifies parsing of IPv6 addresses and Ethernet
MAC addresses in the same way.