Hairpin by enqueueing
Instead of directly calling out to procedures to hairpin packets, treat
them as if they were incoming IPv4 packets by enqueueing them in the
link.  This switch will enable batched lookups, by having a common
continuation for encapsulation and decapsulation after the lookup
results stream in.

Given that the Intel drivers will only pull in a maximum of 128 packets
per breathe() and the link is 256 packets big, we shouldn't see dropped
packets even if we're slammed.  Switch the order of encapsulation and
decapsulation in breathe() to handle hairpinned packets within the same
cycle.