Drop dead incoming packets
If a packet's TTL is or will be 0 after processing, free it.  Fixes a
memory leak.  Thanks to Diego Pino for spotting this one.

Note that this also changes decrement_ttl to never increase the TTL,
which of course makes sense.  We should drop incoming packets with a TTL
of 0 or 1; the former because we shouldn't *receive* packets with TTL=0
but we can't control that, and we shouldn't *send* packets with TTL=0
because of RFC 1122.