lwAFTR uses port=0 for ICMP reply from B4
Just as the lwAFTR uses port=0 when looking up the softwire for an echo
request coming from the internet, a reply from the B4 needs to use the
same port=0 key.  Otherwise any B4 could spoof a reply out to the
internet.  Likewise this also allows pingable psid=0 when there are no
reserved ports.  The successful reply case from an unshared IPv4 is
already tested; this adds a test for a reply whose echo identifier does
correspond to its IPv4 port but which shouldn't be routed because it
should use port=0, and it's not bound to port=0.