RateLimiter leaky=false
This adds a configuration argument to RateLimiter, leaky.  If it's true,
you have the previous mode of operation where incoming traffic that's
above the threshold is dropped.  Otherwise it is left on the incoming
ring buffer.

As a functional change, this new parameter defaults to false, meaning
that users of RateLimiter should update their configurations.