vhost_user: Fix & comment "feature cache"
Added a long description of the "feature cache" and how it supports
reconnecting to a running QEMU.

Added a restriction so that the cache is only considered for the first
feature negotiation performed by a vhost-user app instance. This is to
avoid using the cached features when renegotiating with a guest e.g.
when the guest has switched to a new driver. Not yet tested.