vhost-user: Support operation without MRG_RXBUF.
MRG_RXBUF is enabled by default in the beginning, and QEMU will initially
negotiate a feature set with Snabb NFV that includes MRG_RXBUF.  This
adds a field onto the virtio header, in legacy mode.  However if we
later negotiate to not have MRG_RXBUF, we need to re-set this value to
not have the extra fields.