Indirect access to packet data
Add a pointer to the start of the packet, which is a cursor into the
bytes.  Adjust so that the normal packet data starts 64 bytes into the
packet, giving space for user code to prepend a full IPv6 header.  Adapt
new_packet() to initialize the p.data member, free() to reset the data
member, and adapt prepend, shiftleft, and shiftright to avoid copying if
possible.