Bump packet headroom to 64 bytes.
The headroom is 48 right now which is enough for a 40-byte IPv6 header.
However it's not enough for that and a virtio header, which is 12 or 14
bytes. But, in the lwaftr most packets come from virtio, so there will
be enough -- but, some packets don't, like NDP packets or something.
Better to just reserve a few more bytes so that we know that the virtio
interfaces won't have to memmove.